+++ services/catalog/src/simcore_service_catalog/api/routes/services_resources.py
@@ -5,10 +5,6 @@ from typing import Any, Final, cast
 
 import yaml
 from fastapi import APIRouter, Depends, HTTPException, status
-from models_library.api_schemas_catalog.schemas.constants import (
-    RESPONSE_MODEL_POLICY,
-    SIMCORE_SERVICE_SETTINGS_LABELS,
-)
 from models_library.docker import DockerGenericTag
 from models_library.groups import GroupAtDB
 from models_library.service_settings_labels import (
